Isolate thinking placeholder
